Removing 0993 app, new opt out modal link
