--- src/applications/edu-benefits/0993/edu-benefits-entry.jsx
@@ -1,15 +1,0 @@
-import 'platform/polyfills';
-import '../sass/edu-benefits.scss';
-
-import startApp from 'platform/startup';
-
-import routes from './routes';
-import reducer from './reducers';
-import manifest from './manifest.json';
-
-startApp({
-  url: manifest.rootUrl,
-  reducer,
-  routes,
-  entryName: manifest.entryName,
-});
