--- src/applications/edu-benefits/0993/manifest.json
@@ -1,7 +1,0 @@
-{
-  "appName": "Opt Out of Sharing VA Education Benefits Information",
-  "entryFile": "./edu-benefits-entry.jsx",
-  "entryName": "0993-edu-benefits",
-  "rootUrl": "/education/opt-out-information-sharing/opt-out-form-0993",
-  "productId": "3dbf9046-e929-4c62-b5fe-7789633487aa"
-}
