--- src/applications/edu-benefits/0993/reducers/index.js
@@ -1,6 +1,0 @@
-import formConfig from '../config/form';
-import { createSaveInProgressFormReducer } from 'platform/forms/save-in-progress/reducers';
-
-export default {
-  form: createSaveInProgressFormReducer(formConfig),
-};
