--- src/applications/edu-benefits/0993/tests/e2e/fixtures/mocks/application-submit.json
@@ -1,9 +1,0 @@
-{
-  "data": {
-    "attributes": {
-      "confirmationNumber": "123fake-submission-id-567",
-      "submittedAt": "2016-05-16",
-      "regionalOffice": "Test"
-    }
-  }
-}
