--- src/applications/edu-benefits/0993/tests/schema/maximal-test.json
@@ -1,11 +1,0 @@
-{
-  "data": {
-    "vaFileNumber": "23444444",
-    "claimantFullName": {
-      "first": "claimantFirst",
-      "middle": "claimantMiddle",
-      "last": "claimantLast"
-    },
-    "view:noSSN": true
-  }
-}
