+++ src/applications/edu-benefits/0994/helpers.js
@@ -1,0 +1,10 @@
+import { transformForSubmit } from 'platform/forms-system/src/js/helpers';
+
+export function transform(formConfig, form) {
+  const formData = transformForSubmit(formConfig, form);
+  return JSON.stringify({
+    educationBenefitsClaim: {
+      form: formData,
+    },
+  });
+}
